my_bot_folder/requirements.txt
discord.py>=2.3.2
aiohttp>=3.9.0
aiosqlite>=0.20.0
